Move game and square in separate package
