+++ game/game.go
@@ -4,20 +4,29 @@
  *
  * Contains Game type and methods to interact with it
  */
-package server
+package game
+
+import (
+	// System
+	// Third-party
+	// Project
+	"github.com/BaldaGo/balda-go/conf"
+)
 
 /**
  * @class Game
  * @brief Class, provide information about concrete game
  */
 type Game struct {
-	square *Square ///< Gaming area
+	square Square ///< Gaming area
 }
 
 /**
  * @brief Create a new game
  * @return game Pointer to the created Game object
  */
-func NewGame() *Game {
-	return &Game{square: NewSquare(5)}
+func NewGame(cfg conf.GameConf) Game {
+	return Game{
+		square: NewSquare(cfg.AreaSize),
+	}
 }
