+++ game/square.go
@@ -5,7 +5,7 @@
  * Contans method for find and initialize Square object
  */
 
-package server
+package game
 
 import (
 	// System
@@ -35,8 +35,8 @@ type Square struct {
  *
  * Create new Square and initialize them with random word
  */
-func NewSquare(size int) *Square {
-	area := new(Square)
+func NewSquare(size int) Square {
+	var area Square
 	area.matrix = make([][]rune, size)
 
 	for i := range area.matrix {
@@ -71,7 +71,7 @@ func (area *Square) destructor() {
  * A shallow copy is not suitable for our algorithm,
  * because recursion should work with a copy of our matrix
  */
-func (b Square) deepCopy(a Square) {
+func (b *Square) deepCopy(a Square) {
 	for i := range a.matrix {
 		for j := range a.matrix[i] {
 			b.matrix[i][j] = a.matrix[i][j]
@@ -97,7 +97,7 @@ func (area *Square) addUsedWord(word string) {
  *
  * Find word into list of words
  */
-func (area *Square) wordAlreadyUsed(word []rune) bool {
+func (area Square) wordAlreadyUsed(word []rune) bool {
 	for i := range area.usedWords {
 		if area.usedWords[i] == string(word) {
 			logger.Log.Debugf("Word '%s' used already in this game", string(word))
@@ -135,7 +135,7 @@ func (area Square) PrintUsedWords() {
  * @param[in] y
  * @param[in] symbol
  */
-func (area Square) addSymbol(x int, y int, symbol rune) {
+func (area *Square) addSymbol(x int, y int, symbol rune) {
 	area.matrix[x][y] = symbol
 	logger.Log.Debugf("New symbol '%c' added", symbol)
 }
@@ -214,7 +214,7 @@ func (area *Square) CheckWord(x int, y int, symbol rune, word []rune) bool {
 		for j := range area.matrix[i] {
 			if tempArea.matrix[i][j] == rune(word[0]) &&
 				(tempArea.findFull(x, y, i, j, word, false) != 0) {
-				area.deepCopy(*tempArea)
+				area.deepCopy(tempArea)
 				area.addUsedWord(string(word))
 				logger.Log.Debugf("New word '%s' added", string(word))
 				return true
